+++ packages/editor/src/components/editor-global-keyboard-shortcuts/index.js
@@ -152,7 +152,6 @@ export default compose( [
 			getBlockOrder,
 			getMultiSelectedBlockClientIds,
 			hasMultiSelection,
-			isEditedPostDirty,
 			getBlockRootClientId,
 			getTemplateLock,
 			getSelectedBlockClientId,
@@ -167,11 +166,10 @@ export default compose( [
 				selectedBlockClientIds,
 				( clientId ) => !! getTemplateLock( getBlockRootClientId( clientId ) )
 			),
-			isDirty: isEditedPostDirty(),
 			selectedBlockClientIds,
 		};
 	} ),
-	withDispatch( ( dispatch, ownProps ) => {
+	withDispatch( ( dispatch, ownProps, { select } ) => {
 		const {
 			clearSelectedBlock,
 			multiSelect,
@@ -185,11 +183,11 @@ export default compose( [
 			onSave() {
 				// TODO: This should be handled in the `savePost` effect in
 				// considering `isSaveable`. See note on `isEditedPostSaveable`
-				// selector about dirtiness and meta-boxes. When removing, also
-				// remember to remove `isDirty` prop passing from `withSelect`.
+				// selector about dirtiness and meta-boxes.
 				//
 				// See: `isEditedPostSaveable`
-				if ( ! ownProps.isDirty ) {
+				const { isEditedPostDirty } = select( 'core/editor' );
+				if ( ! isEditedPostDirty() ) {
 					return;
 				}